itest: move helper functions into one file
This commit creates the file utils.go to hold the commonly used
functions for tests.
